requirements.txt
aldryn-bootstrap3==1.3.0
aldryn-style==1.1.4
Django==2.1.9
django-appconf==1.0.3
django-classy-tags==0.9.0
django-cms==3.6.0
django-filer==1.5.0
django-formtools==2.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-bootstrap4==1.3.1
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-icon==1.4.0
djangocms-link==2.4.0
djangocms-modules==0.2.0
djangocms-picture==2.2.1
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-transfer==0.2.0
djangocms-video==2.1.1
easy-thumbnails==2.6
html5lib==1.0.1
Pillow==6.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
webencodings==0.5.1
